move hooks to its own package
