+++ internal/ent/hooks/doc.go
@@ -1,0 +1,2 @@
+// Package hooks is middleware to alter the graphql mutation
+package hooks
