+++ internal/ent/hooks/errors.go
@@ -1,0 +1,10 @@
+package hooks
+
+import (
+	"errors"
+)
+
+var (
+	// ErrInternalServerError is returned when an internal error occurs.
+	ErrInternalServerError = errors.New("internal server error")
+)
