+++ internal/ent/schema/errors.go
@@ -10,7 +10,4 @@ var (
 
 	// ErrContainsSpaces is returned when field contains spaces
 	ErrContainsSpaces = errors.New("field should not contain spaces")
-
-	// ErrInternalServerError is returned when an internal error occurs.
-	ErrInternalServerError = errors.New("internal server error")
 )
